+++ go-app/frontend/src/tabs/calibration/RemapNotchesModal.tsx
@@ -1,30 +1,31 @@
 import { main } from "../../../wailsjs/go/models";
 import { CalibrationModalForm } from "./CalibrationModalForm";
 import { Modal, ModalContentProps } from "../../components";
+import { RemapNotchesModalForm } from "./RemapNotchesModalForm";
 
 type Props = {
   controller: main.Interop_GenericController | null;
   onClose: () => void;
 };
 
-const CalibrationModalContent = ({
+const RemapNotchesModalContent = ({
   openState: controller,
   onClose,
 }: ModalContentProps<main.Interop_GenericController>) => {
   return (
-    <CalibrationModalForm
+    <RemapNotchesModalForm
       controller={controller}
       onClose={onClose}
     />
   );
 };
 
-export const CalibrationModal = ({ controller, onClose }: Props) => {
+export const RemapNotchesModal = ({ controller, onClose }: Props) => {
   return (
     <Modal
       openState={controller ?? false}
       onClose={onClose}
-      Component={CalibrationModalContent}
+      Component={RemapNotchesModalContent}
     />
   );
 };