+++ go-app/profile_runner/api_controller.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"fmt"
 	"math"
+	"regexp"
+	"strings"
 	"sync"
 	"time"
 	"tsw_controller_app/logger"
@@ -35,14 +37,70 @@ type ApiController_ControlTargets struct {
 	mutex sync.RWMutex
 }
 
+type ApiController_ActiveCab struct {
+	updatedAt *time.Time
+	Front     bool
+	Back      bool
+}
+
 type ApiController struct {
-	API            *tswapi.TSWAPI
+	API            tswapi.ITSWAPI
 	ControlChannel chan ApiController_Command
+	ActiveCab      ApiController_ActiveCab
 	interacting    ApiController_Interacting
 }
 
 func (c *ApiController_Command) ToString() string {
 	return fmt.Sprintf("api_control_command:%s:%f", c.Controls, c.InputValue)
+}
+
+func (s *ApiController_ActiveCab) Update(api tswapi.ITSWAPI) error {
+	/* update at most once every 2 seconds */
+	now := time.Now()
+	should_update := s.updatedAt == nil || now.Sub(*s.updatedAt).Seconds() > 2.0
+	if should_update {
+		cab, err := api.GetActiveCab()
+		s.Front = cab.Front
+		s.Back = cab.Back
+		s.updatedAt = &now
+		return err
+	}
+	return nil
+}
+
+func (controller *ApiController) formatControlName(control string) (string, error) {
+	re := regexp.MustCompile(`\{SIDE\}|\{SIDE:[^:\}]+:[^:\}]+\}`)
+
+	side_replacement_failed := false
+	formatted_control := re.ReplaceAllStringFunc(control, func(match string) string {
+		err := controller.ActiveCab.Update(controller.API)
+		if err != nil {
+			logger.Logger.Error("[App::ApiController] Could not update active cab: %e", err)
+		}
+
+		match_parts := strings.Split(match[1:len(match)-1], ":") /* split and remove leading and trailing {} */
+		front_value := "F"
+		back_value := "B"
+		if len(match_parts) == 3 {
+			front_value = match_parts[1]
+			back_value = match_parts[2]
+		}
+		if controller.ActiveCab.Back {
+			return back_value
+		}
+		if controller.ActiveCab.Front {
+			return front_value
+		}
+
+		side_replacement_failed = true
+		return match
+	})
+
+	if side_replacement_failed {
+		return formatted_control, fmt.Errorf("Could not replace side placeholder due to missing active cab")
+	}
+
+	return formatted_control, nil
 }
 
 func (controller *ApiController) StartInteractingIfNotAlready(ctx context.Context, control string) error {
@@ -134,7 +192,8 @@ func (controller *ApiController) ProcessPendingControlCommand(ctx context.Contex
 	defer controller.clearControlTargetCommand(command)
 
 	/* we're just silently ignoring this error here and starting from a default of 0.0f on failure which is acceptable in most cases */
-	current_value, _ := controller.API.GetInputValue(command.Controls)
+	current_value := 0.0
+	current_value, _ = controller.API.GetInputValue(command.Controls)
 	target_value_diff := math.Abs(current_value - command.InputValue)
 	/* no-op if the value is already the same */
 	if math_utils.IsWithinMarginOfError(current_value, command.InputValue) {
@@ -189,15 +248,26 @@ func (controller *ApiController) ProcessPendingControlStates(ctx context.Context
 * - Assigns incoming command as the target command
  */
 func (controller *ApiController) ProcessControlCommand(ctx context.Context, command ApiController_Command) error {
-	if err := controller.StartInteractingIfNotAlready(ctx, command.Controls); err != nil {
+	control, err := controller.formatControlName(command.Controls)
+
+	if err != nil {
+		return err
+	}
+
+	if err := controller.StartInteractingIfNotAlready(ctx, control); err != nil {
 		return err
 	}
 
 	controller.interacting.mutex.Lock()
 	defer controller.interacting.mutex.Unlock()
-	controlstate := controller.interacting.controls[command.Controls]
-	controlstate.TargetCommand = &command
-	controller.interacting.controls[command.Controls] = controlstate
+	controlstate := controller.interacting.controls[control]
+	controlstate.TargetCommand = &ApiController_Command{
+		Controls:      control,
+		InputValue:    command.InputValue,
+		Hold:          command.Hold,
+		MaxChangeRate: command.MaxChangeRate,
+	}
+	controller.interacting.controls[control] = controlstate
 	return nil
 }
 
@@ -222,10 +292,15 @@ func (controller *ApiController) Run(ctx context.Context) func() {
 	return cancel
 }
 
-func NewAPIController(twapi *tswapi.TSWAPI) *ApiController {
+func NewAPIController(twapi tswapi.ITSWAPI) *ApiController {
 	controller := ApiController{
 		API:            twapi,
 		ControlChannel: make(chan ApiController_Command, API_CONTROLLER_QUEUE_BUFFER_SIZE),
+		ActiveCab: ApiController_ActiveCab{
+			updatedAt: nil,
+			Front:     false,
+			Back:      false,
+		},
 		interacting: ApiController_Interacting{
 			mutex:    sync.RWMutex{},
 			controls: map[string]ApiController_Interacting_Control{},